// Angular/src/app/Enpoints.g.ts
export class VehicleDto implements IVehicleDto {
    id!: string;
    name!: string;
    year!: number;
    cost!: number;

    constructor(data?: IVehicleDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.year = _data["year"];
            this.cost = _data["cost"];
        }
    }

    static fromJS(data: any): VehicleDto {
        data = typeof data === 'object' ? data : {};
        let result = new VehicleDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["year"] = this.year;
        data["cost"] = this.cost;
        return data;
    }
}

export interface IVehicleDto {
    id: string;
    name: string;
    year: number;
    cost: number;
}
// ...
export class UpsertVehicleDto implements IUpsertVehicleDto {
    id!: string;
    name!: string;
    year!: number;
    cost!: number;

    constructor(data?: IUpsertVehicleDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.year = _data["year"];
            this.cost = _data["cost"];
        }
    }

    static fromJS(data: any): UpsertVehicleDto {
        data = typeof data === 'object' ? data : {};
        let result = new UpsertVehicleDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["year"] = this.year;
        data["cost"] = this.cost;
        return data;
    }
}

export interface IUpsertVehicleDto {
    id: string;
    name: string;
    year: number;
    cost: number;
}